move linear algebra utils to utils.h
