+++ src/quiz/ransac/ransac2d.cpp
@@ -12,6 +12,8 @@
 // using templates for processPointClouds so also include .cpp to help linker
 #include "../../processPointClouds.cpp"
 
+#include "utils.h"
+
 pcl::PointCloud<pcl::PointXYZ>::Ptr CreateData()
 {
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud(new pcl::PointCloud<pcl::PointXYZ>());
@@ -67,36 +69,6 @@ pcl::visualization::PCLVisualizer::Ptr initScene()
 }
 
 
-/* Return wether the pointA and pointB are overlapping */
-bool overlapping(pcl::PointXYZ pointA, pcl::PointXYZ pointB) {
-	return (pointA.x == pointB.x) && (pointA.y == pointB.y) && (pointA.z == pointB.z);
-
-}
-
-/* Compute the unit vector of the AB line */
-pcl::PointXYZ vector(pcl::PointXYZ pointA, pcl::PointXYZ pointB) {
-	return pcl::PointXYZ((pointB.x - pointA.x), (pointB.y - pointA.y), (pointB.z - pointA.z));
-}
-
-float norm(pcl::PointXYZ vect) {
-	return (float) sqrt(pow(vect.x, 2) + pow(vect.y, 2) + pow(vect.z, 2));
-}
-
-pcl::PointXYZ unit( pcl::PointXYZ vect) {
-	auto norm_ = norm(vect);
-	return pcl::PointXYZ(vect.x/norm_, vect.y/norm_, vect.z/norm_);
-}
-
-float scalar_product(pcl::PointXYZ pointB, pcl::PointXYZ pointA) {
-	return (float) ((pointA.x * pointB.x) + (pointA.y * pointB.y) + (pointA.z * pointB.z));
-}
-
-
-float distance_to_line(pcl::PointXYZ point, pcl::PointXYZ linePointA, pcl::PointXYZ vectorAB) {
-	auto vectorAX = vector(linePointA, point);
-	return sqrt(1-pow(scalar_product(unit(vectorAX), unit(vectorAB)),2)) * norm(vectorAX);
-}
-
 std::tuple<int, int,  std::unordered_set<int>> Ransac(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud, int maxIterations, float distanceTol)
 {
 
@@ -135,7 +107,7 @@ std::tuple<int, int,  std::unordered_set<int>> Ransac(pcl::PointCloud<pcl::Point
 		// Measure distance between every point and fitted line
 		auto vectorAB = vector(pointA, pointB);
 
-		std::cout << "Distance B to AB (should be 0)" << distance_to_line(pointB, pointA, vectorAB) << std::endl;
+		// std::cout << "Distance B to AB (should be 0)" << distance_to_line(pointB, pointA, vectorAB) << std::endl;
 		for (int index = 0; index < cloud->points.size(); index++) {
 			// If distance is smaller than threshold count it as inlier
 			if (inliers.count(index)>0)
