+++ src/quiz/ransac/utils.h
@@ -1,0 +1,31 @@
+#include <pcl/common/common.h>
+
+/* Return wether the pointA and pointB are overlapping */
+bool overlapping(pcl::PointXYZ pointA, pcl::PointXYZ pointB) {
+	return (pointA.x == pointB.x) && (pointA.y == pointB.y) && (pointA.z == pointB.z);
+
+}
+
+/* Compute the unit vector of the AB line */
+pcl::PointXYZ vector(pcl::PointXYZ pointA, pcl::PointXYZ pointB) {
+	return pcl::PointXYZ((pointB.x - pointA.x), (pointB.y - pointA.y), (pointB.z - pointA.z));
+}
+
+float norm(pcl::PointXYZ vect) {
+	return (float) sqrt(pow(vect.x, 2) + pow(vect.y, 2) + pow(vect.z, 2));
+}
+
+pcl::PointXYZ unit( pcl::PointXYZ vect) {
+	auto norm_ = norm(vect);
+	return pcl::PointXYZ(vect.x/norm_, vect.y/norm_, vect.z/norm_);
+}
+
+float scalar_product(pcl::PointXYZ pointB, pcl::PointXYZ pointA) {
+	return (float) ((pointA.x * pointB.x) + (pointA.y * pointB.y) + (pointA.z * pointB.z));
+}
+
+
+float distance_to_line(pcl::PointXYZ point, pcl::PointXYZ linePointA, pcl::PointXYZ vectorAB) {
+	auto vectorAX = vector(linePointA, point);
+	return sqrt(1-pow(scalar_product(unit(vectorAX), unit(vectorAB)),2)) * norm(vectorAX);
+}
